move makefile & modules around
